Modify code

+++ 3_Implementation/inc/BMI.h
@@ -1,10 +1,8 @@
 #ifndef __BMI_H__
 #define __BMI_H__
 #include<stdio.h>
-float Kg_Meter(int,float);
-float Kg_Cm(int,float);
-float Lbs_Cm(int,float);
-float Lbs_Meter(int,float);
-
-
+float Kg_Meter(float,float);
+float Kg_Cm(float,float);
+float Lbs_Cm(float,float);
+float Lbs_Meter(float,float);
 #endif
